feature: check device location and get current location
Continue the analysis about location by first checking if the device location is enabled or not, and provide the user indications and prompts to activate it.
Then, get the current location of the user and, if successfully retrieved, decode the location into a human readable address.
Exploit the use of live data variables declared in the view model class.

Integrates: #9