refactor: provide necessary connections to the current code
Provide comments to the code when required.
Change the name of some functions related to view model flags in order to have coherence in the project.
Set the app bar title when navigating to the representative fragment.
Add the network status whenever there is an HTTP call to the civics API.

Integrates: #9