refactor: use IO dispatcher in coroutine scope
When interacting with the database, do not use the main dispatcher to resolve the coroutine but switch to the IO dispatcher, since Android must read from or write to the I/O disk.
Then, update the fabStatus value in VoterInfoViewModel according to the fact that the election is saved or not.

Integrates: #19