refactor: change fragment label argument and deactivate flags properly
Change the VoterInfoFragment label inside the navigation graph, because the previous value was unattainable: instead of trying to retrieve the election name from the election argument, pass the election name as an additional argument, that will be only used as the fragment app bar title.
Then, put the flags deactivation for the voter info and ballot info in the right place.