refactor: change control of the motion layout state among fragment re-creation
Upon configuration changes or other situations where fragments must be destroyed and re-created, handle correctly the restore of the motion layout state.
Add also a check on the form variables as soon as the "Find My Representatives button is clicked". Show an alert dialog if one of the entries is not filled correctly.