refactor: use LiveData implementation for restoring the list
Instead of saving inside the SavedStateHandle class the data and to push that data into the LiveData value, use the LiveData implementation of SavedStateHandle in order to keep the live data variable updated.
Moreover do not use the init and the onCleared method.
Change some string resources giving more detailed names.